fix: migrate kit to the hardened + extended arvel API; native builder/DDL
Consume the new arvel capabilities and fix every API break the parity round and
the follow-on framework work introduced.

- HTTP client: response.status_code/.content -> response.ok()/.content(); the
  gateway charge + image download were silently broken.
- auth: password reset via PasswordBroker (+ email in ResetPasswordIn, a
  PasswordResetRequested listener, password_reset_tokens migration); email verify
  binds the user id (VerifyEmailIn.id); remember_tokens + api_tokens.last_used_at
  migrations (the missing last_used_at 500'd every authenticated request).
- search: Model.search() -> .get(); cache: lock(seconds=) + LockAcquireFailed.
- framework-native, no raw SQLAlchemy: deal sell-through via DB.table() with a
  qualified-column join filter; visibility refresh via DB.refresh_materialized_view();
  DB.dialect(); request.body() for webhook HMAC.
- security: bind-param the deals aggregate, convert type-narrowing asserts to real
  guards; [tool.bandit] + hook scoped to app.

Verified: 136 served-path unit tests green, mypy/pyright strict + bandit clean,
migrate:fresh applies all migrations, real seed attaches product media.

+++ app/services/coupon_service.py
@@ -56,8 +56,8 @@ def discount_cents(coupon: Coupon, subtotal_cents: int) -> int:
         coupon.type if isinstance(coupon.type, CouponType) else CouponType(coupon.type)
     )
     if kind is CouponType.PERCENT:
-        return min(subtotal_cents * coupon.value // 100, subtotal_cents)
-    return min(coupon.value, subtotal_cents)
+        return int(min(subtotal_cents * coupon.value // 100, subtotal_cents))
+    return int(min(coupon.value, subtotal_cents))
 
 
 def normalize_code(code: str) -> str:
